dynamic versioning
updated my template with a few things I have been using lately. But mainly
rewrote the setup.py file to allow for dynamic versioning. Now, I have done this
in hackish ways before, but this time I think I have come up with a versatile
way that supports both "derived" dynamic versioning as well as allows for CI/CD
processes where a human in the loop wishes to approve (and provide the version
number) publication of releases.

For "internal" purposes, where a proprietary PyPI is being used, every commit
can be an internal release and versions can be dynamically bumped. However, this
process has a flaw - one has to remember to change the version control file,
which itself isn't actually a change to the code, and if one forgets to make
that change along with actual code, one might be forced to make a commit that is
*merely* making the change to the version control file. Ick.

This new way allows one to provide either an explicit version (`--new-version`)
or a directive to bump a portion of the semantic version
(`--version-bump=["major", "minor", "update"]`) to the `sdist` subcommand. If a
release builds are manually kicked off, this value can be provided by the human
in the loop. Meanwhile, if internal release builds are dynamically kicked off,
either a passed variable can populate one of those two values, or it can fall
back on incrementing the "update" portion of the version by default. This seems
the best of both worlds without BS commits. Sure, an internal server *could* end
up with a version nobody wanted, but a package is easily deleted from an
internal PyPI, and a manual build can be kicked off to provide that commit with
the appropriate version.