default version is dev / current release
I decided that the default behavior of sdist should be to create a "dev"
version, which is usually represented by a number of commits *forward* from the
latest MAJOR release (I chose major, as I think it makes the most sense - using
--new-version can always be used to override things, even to create dev
versions). So for instance, if `git describe --long` were to return
2.4.9-12-gd2d97f8, I note that the last major release was 2.0.0. I also note
that there have been commits since the last release (regardless of level). I
perform a `git rev-list` to determine how many commits there have been since the
2.0.0 release (in this case `git rev-list 2.0.0..HEAD --count`). Armed with that
count, I then assemble the dev release as such: 3.0.0.dev{count}.

One the one caveat is when `git describe` returns with 0 commits since the last
release. In this case, I assume we want to package *this* release, and take the
version wholesale.

This way, the default use of `python setup.py sdist` will either produce a
package of the version described by this very commit, or it will create a
development release based on how many changes have been made since the last
major release, creating a version that follows PEP 440, where a development
release always presceeds the actual release but is after what is presumed to be
the latest release (2.4.9 < 3.0.0.dev12 < 3.0.0).